pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "network-security"
version = "0.1.0"
requires-python = ">=3.8"

authors = [
  { name="Sourabh Biradar", email="sourabhbiradar@outlook.com" },
]

dependencies = [
    "numpy",
    "pandas",
    "pymongo>=4.0",
    "python-dotenv",
    "certifi"
]

[tool.setuptools.packages.find]
where = ["src"]